updated code v1.0.0.4

+++ src/main/java/dev/anhcraft/yumn/heightmaps/HeightMapManager.java
@@ -10,10 +10,10 @@ public class HeightMapManager {
     private final static Logger LOGGER = new Logger("HeightMapManager");
     private final Map<UUID, WorldHeightMap> data = new ConcurrentHashMap<>();
 
-    public WorldHeightMap request(UUID worldId){
+    public WorldHeightMap request(UUID worldId) {
         return data.compute(worldId, (uuid, m) -> {
-            if(m == null) {
-                LOGGER.log("Heightmap for #%s not found! Creating new one...", worldId.toString());
+            if (m == null) {
+                LOGGER.logf("Heightmap for #%s not found! Creating new one...", worldId.toString());
                 return new WorldHeightMap();
             } else {
                 return m;
@@ -25,7 +25,7 @@ public class HeightMapManager {
      * @deprecated Internal uses only!
      */
     @Deprecated
-    public void clean(){
+    public void clean() {
         data.values().forEach(WorldHeightMap::clean);
     }
 }
